Main

Description

This trusted module defines the program's main function. It's Unsafe, and so may not be used by Safe modules.

Documentation

main :: IO ()

Program's entry point. When main is run, it matches the program's command line arguments against the following patterns, issuing an error message if pattern matching fails:

["server", port] Starts the server running on the specified port.

["client", host, port] Starts an instance of the client, specifying the host and port to connect to.